+++ src/aws/secrets_helpers.go
@@ -54,8 +54,8 @@ func ReadSecrets(awsConfig types.AwsConfig) (types.Secrets, error) {
 // MergeAndWriteSecrets merges two secret maps and writes them to s3
 // Overwrites existingSecrets's values if the are conflicting keys
 func MergeAndWriteSecrets(existingSecrets, newSecrets types.Secrets, awsConfig types.AwsConfig) error {
-	secrets := util.Merge(existingSecrets, newSecrets)
-	return writeSecrets(secrets, awsConfig)
+	util.Merge(existingSecrets, newSecrets)
+	return writeSecrets(existingSecrets, awsConfig)
 }
 
 // DeleteSecrets deletes a list of secrets provided their keys. Ignores them if they don't exist
